[FIX] Fixed some bugs : navbar, user view, home view, landing page - Added titles to the badges in the navigation bar - The amount of carbon saved and the score in the leaderboard is consistent - Fixed other minnors bugs (home view, user view, typos)

+++ back-end/erasmail/emails/managers.py
@@ -56,8 +56,8 @@ class EmailStatsQuerySet(models.QuerySet):
                 F("shared_badges") * 1000
                 + F("shared_stats") * 1000
                 + F("user__connected_count") * 25
-                + (F("saved_carbon")), # the amount of CO2 saved in kg
-                output_field=IntegerField(),
+                + F("saved_carbon"), # the amount of CO2 saved in kg
+                output_field=FloatField(),
             ),
         )
